Add missing docs, refactor
 Changes to be committed:
	deleted:    src/main/java/org/dimdev/vanillafix/client/VanillaFixClient.java
	modified:   src/main/java/org/dimdev/vanillafix/client/modsupport/ModMenuSupport.java
	renamed:    src/main/java/org/dimdev/vanillafix/profiler/mixins/KeyboardMixin.java -> src/main/java/org/dimdev/vanillafix/profiler/mixins/client/KeyboardMixin.java
	renamed:    src/main/java/org/dimdev/vanillafix/profiler/mixins/MinecraftClientMixin.java -> src/main/java/org/dimdev/vanillafix/profiler/mixins/client/MinecraftClientMixin.java
	renamed:    src/main/java/org/dimdev/vanillafix/profiler/mixins/WorldRendererMixin.java -> src/main/java/org/dimdev/vanillafix/profiler/mixins/client/WorldRendererMixin.java
	modified:   src/main/java/org/dimdev/vanillafix/textures/mixins/SpriteAtlasTextureMixin.java
	modified:   src/main/resources/vanillafix.profiler.mixins.json
