--- src/main/java/org/dimdev/vanillafix/client/VanillaFixClient.java
@@ -1,10 +1,0 @@
-package org.dimdev.vanillafix.client;
-
-import net.fabricmc.api.ClientModInitializer;
-
-// TODO: put something here
-public class VanillaFixClient implements ClientModInitializer {
-    @Override
-    public void onInitializeClient() {
-    }
-}
