+++ src/main/java/org/dimdev/vanillafix/profiler/mixins/client/KeyboardMixin.java
@@ -1,4 +1,4 @@
-package org.dimdev.vanillafix.profiler.mixins;
+package org.dimdev.vanillafix.profiler.mixins.client;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -8,7 +8,6 @@ import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
 import net.minecraft.client.Keyboard;
 import net.minecraft.client.gui.hud.ChatHud;
-import net.minecraft.text.TranslatableText;
 
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
@@ -21,6 +20,7 @@ public class KeyboardMixin {
      * */
     @Inject(method = "processF3", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/hud/ChatHud;addMessage(Lnet/minecraft/text/Text;)V", ordinal = 9), locals = LocalCapture.CAPTURE_FAILHARD)
     private void addF3SHelpMessage(int key, CallbackInfoReturnable<Boolean> cir, ChatHud chatGui) {
-        chatGui.addMessage(new TranslatableText("vanillafix.debug.switch_profiler.help"));
+        // TODO
+//        chatGui.addMessage(new TranslatableText("vanillafix.debug.switch_profiler.help"));
     }
 }
